+++ native/src/lib.rs
@@ -1,3 +1,7 @@
 pub mod api;
+pub mod domain;
+pub mod error;
+pub mod executor;
+pub mod service;
 #[rustfmt::skip]
 mod frb_generated; /* AUTO INJECTED BY flutter_rust_bridge. This line may not be accurate, and you can change it according to your needs. */